requirements.txt
chatterbox-tts==0.1.6
torchvision==0.21.0
python-dotenv==1.2.1
runpod==1.8.1
faster-whisper==1.2.1
opencv-python==4.11.0.86
transformers==4.46.3
accelerate==0.26.0